refactor: move generatePath and types into own files
