+++ types.ts
@@ -1,0 +1,9 @@
+export type Hexagon = [number, number];
+
+export type Level = {
+  hexagonList: Hexagon[];
+  startHexagons: Hexagon[];
+  endHexagons: Hexagon[];
+  minPathLenght: number;
+  maxPathLength: number;
+};
